Move secrets tests to manager_rest/tests/
No reason for these tests to be inte-tests! They work just fine as
regular restservice tests - I pretty much just copied them verbatim :)

(almost - I replaced `self.assertX` with just pytest `assert`, and
also changed assertRaises to the `with` syntax)
